Display none for print class
Many components are hidden in the printable version due to being
interactive or simply irrelevant to a user printing a page.
This is currently done in separate print stylesheets.

Thanks to the adoption of a dedicated govuk-!-display-none-print
class in govuk-frontend, we can remove display:none from many
components' print stylesheets and use the special class instead.